+++ android/app/src/main/java/com/formatforge/app/MainActivity.java
@@ -9,6 +9,7 @@ public class MainActivity extends BridgeActivity {
     public void onCreate(Bundle savedInstanceState) {
         registerPlugin(KugouPlugin.class);
         registerPlugin(SaveFilePlugin.class);
+        registerPlugin(AppUpdatePlugin.class);
         super.onCreate(savedInstanceState);
     }
 }